SBERDOMA-431 splitted ContentEditor module into files
