+++ apps/condo/domains/ticket/components/BaseTicketForm/index.tsx
@@ -17,7 +17,7 @@ import { useTicketValidations } from './useTicketValidations'
 import { FrontLayerContainer } from '@condo/domains/common/components/FrontLayerContainer'
 import { useMultipleFileUploadHook } from '@condo/domains/common/components/MultipleFileUpload'
 import { TicketFile, ITicketFileUIState } from '@condo/domains/ticket/utils/clientSchema'
-import { useContactsEditorHook } from '@condo/domains/contact/components/ContactsEditor'
+import { useContactsEditorHook } from '@condo/domains/contact/components/ContactsEditor/useContactsEditorHook'
 import { useOrganization } from '@core/next/organization'
